+++ eyecite/helpers.py
@@ -363,8 +363,7 @@ def _process_case_name(
         else:
             plaintiff, defendant = "", splits[0]
         plaintiff = plaintiff.strip(f"{whitespace},(")
-        clean_plaintiff = re.sub(r"\b[a-z]\w*\b", "", plaintiff)
-        plaintiff = strip_stop_words(clean_plaintiff)
+        plaintiff = strip_stop_words(plaintiff)
         citation.metadata.plaintiff = plaintiff
     else:
         defendant = candidate_case_name